Lockscreen clock styles: Move everything to the keyguard [1/2]
- Use XML attributes to keep redundancy down. Same result, less code :-)
