+++ core/res/res/values/custom_colors.xml
@@ -72,11 +72,5 @@
     <!-- Fastscroller -->
     <color name="fastscroller_text_color">#FFFFFFFF</color>
 
-    <!-- Custom Analog clock -->
-    <color name="analog_clock_border_color">#212121</color>
-    <color name="analog_clock_face_color">#FFFFFF</color>
-    <color name="analog_clock_hand_hour_color">@*android:color/accent_device_default_light</color>
-    <color name="analog_clock_hand_minute_color">#212121</color>
-
 </resources>
