+++ core/res/res/values/ex_symbols.xml
@@ -161,14 +161,4 @@
     <!-- Pocket bridge -->
     <java-symbol type="string" name="config_pocketBridgeSysfsInpocket" />
 
-  <!-- Custom Analog clock -->
-  <java-symbol type="drawable" name="custom_clock_dial" />
-  <java-symbol type="drawable" name="custom_clock_dial_ambient" />
-  <java-symbol type="drawable" name="custom_clock_hand_hour" />
-  <java-symbol type="drawable" name="custom_clock_hand_minute" />
-  <java-symbol type="color" name="analog_clock_border_color" />
-  <java-symbol type="color" name="analog_clock_face_color" />
-  <java-symbol type="color" name="analog_clock_hand_hour_color" />
-  <java-symbol type="color" name="analog_clock_hand_minute_color" />
-
 </resources>
