+++ packages/SystemUI/res-keyguard/drawable/custom_clock_dial.xml
