+++ packages/SystemUI/res-keyguard/drawable/custom_clock_hand_hour.xml
