+++ packages/SystemUI/res-keyguard/drawable/custom_clock_hand_minute.xml
